+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Consolidate the mqttOptions in the layer sensor with the documentation (services.json.md).
 - In config.json.md, the attribute `isActive` was changed to `active` in Portalconfig.menu.tool.gfi.
 - External WMS Layers with version lower than 1.3.0 can also be imported.
+- Legends are only printed from layers that are switched visible.
 
 ### Deprecated